+++ tools/ice-devtools/config/getBabelConfig.js
@@ -48,16 +48,7 @@ module.exports = (buildConfig = {}) => {
       ],
       ['babel-plugin-import', { libraryName: '@alife/next' }, '@alife/next'],
       ['babel-plugin-import', { libraryName: '@alifd/next' }, '@alifd/next'],
-      ['babel-plugin-import', { libraryName: 'antd', style: true }, 'antd'],
-      [
-        '@babel/plugin-transform-runtime',
-        {
-          "corejs": false,
-          "helpers": false,
-          "regenerator": true,
-          "useESModules": false
-        }
-      ]
+      ['babel-plugin-import', { libraryName: 'antd', style: true }, 'antd']
     ]),
   };
 };